chrome.bookmarks.update is asynchronous!!
Only log the update action and present the user with an undo
notification after the update completes though a callback.

If not then it is up to the speed with which the user clicks the
notification and race conditions whether the update is performed first
and then the rollback action or the other way round.